+++ resources/views/places/edit.blade.php
@@ -250,7 +250,7 @@
                             <label for="url" class="col-md-3 col-form-label text-md-right">Adresse web</label>
 
                             <div class="col-md-9">
-                                <input id="url" type="text" class="form-control @error('url') is-invalid @enderror" name="url" value="{{ $place->url }}" autocomplete="url" autofocus>
+                                <input id="url" type="text" class="form-control @error('url') is-invalid @enderror" name="url" value="{{ $place->url }}" autocomplete="url" placeholder="Ex.: https://solutionlocale.ca"  autofocus>
 
                                 @error('url')
                                     <span class="invalid-feedback" role="alert">
@@ -259,6 +259,27 @@
                                 @enderror
                             </div>
                         </div>
+
+                        <div class="form-group row">
+                                <label for="facebook_url" class="col-md-3 col-form-label text-md-right">
+                                    Adresse de page Facebook
+                                </label>
+
+                                <div class="col-md-9">
+                                    <input id="facebook_url" type="text"
+                                           class="form-control @error('facebook_url') is-invalid @enderror"
+                                           name="facebook_url"
+                                           value="{{ $place->facebook_url }}"
+                                           placeholder="Ex.: https://www.facebook.com/Solutionlocale/" 
+                                           autocomplete="facebook_url" autofocus>
+
+                                    @error('facebook_url')
+                                        <span class="invalid-feedback" role="alert">
+                                            <strong>{!! $message !!}</strong>
+                                        </span>
+                                    @enderror
+                                </div>
+                            </div>
 
                         <div class="form-group row">
                             <label for="deliveryZone" class="col-md-3 col-form-label text-md-right">Secteur desservi pour la livraison</label>